refactor(types): UI metadata 抽出 + PrimaryKeyConstraint TS sync (#1186 Phase 2-A)
main 厳格レビュー #1186 Phase 2 (TS types/action.ts 完全整理) の最初の段階 (Phase 2-A) を実装。
+ 副次的に PR #1198 (#1185 提案 C) で欠けていた v3 TS 型 sync 完了。

## Phase 2-A: UI metadata 抽出

action.ts (343 行) に同居していた UI 表示メタデータ (LABELS / ICONS / COLORS / VALUES)
を専用モジュール `frontend/src/utils/processFlowMetadata.ts` に分離。

責務分離:
- types/action.ts → schema 型 alias (v3 re-export + AnyRecord 互換層)
- utils/processFlowMetadata.ts → UI 表示専用メタデータ (frontend 固有、schema 概念ではない)

抽出した定数 (11 個):
- ACTION_TRIGGER_LABELS / PROCESS_FLOW_TYPE_LABELS / PROCESS_FLOW_TYPE_ICONS
- STEP_TYPE_LABELS / STEP_TYPE_ICONS / STEP_TYPE_COLORS
- EXTERNAL_CALL_OUTCOME_VALUES / WORKFLOW_PATTERN_VALUES / WORKFLOW_PATTERN_LABELS
- DB_OPERATION_LABELS / STEP_TEMPLATES
- 関連型: ExternalCallOutcome / StepTemplate

action.ts は backward compat re-export 層として機能 (旧 consumer の import は不変):
```ts
// 旧来 (引き続き動作)
import { STEP_TYPE_LABELS } from "@/types/action";
// 新規推奨
import { STEP_TYPE_LABELS } from "@/utils/processFlowMetadata";
```

## PrimaryKeyConstraint TS sync (PR #1198 follow-up)

PR #1198 で schema に新設した PrimaryKeyConstraint が v3 TS 型に未反映で、
tableValidation.ts:18 が型エラーになっていた問題を解消。

- frontend/src/types/v3/table.ts: PrimaryKeyConstraint interface 追加 + Constraint union に含める
- frontend/src/components/table/ConstraintsTab.tsx: kindBadge / constraintSummary に primaryKey case 追加
- frontend/src/utils/ddlGenerator.ts: constraintToDdl に PRIMARY KEY DDL 生成追加 (composite PK 対応)

ConstraintsTab.tsx の local ConstraintKind type (UI 選択肢) は意図的に primaryKey を含めず、
旧来通り Column.primaryKey で表現する UI フロー維持 (PrimaryKeyConstraint は schema/AI 経由のみ
作成可能、UI 編集対応は別 ISSUE)。

## Test plan

- [x] frontend build (tsc + vite) → OK
- [x] frontend test → 2346 / 2346 pass
- [x] backend test → 516 / 516 pass
- [x] validate:samples (retail) → 7 / 7 flows pass

## 残作業 (#1186)

- Phase 2-B: consumer の `import from "@/types/action"` → `@/types/v3` / `@/utils/processFlowMetadata` へ段階移行 (99 file、複数 PR 想定)
- Phase 3: action.ts 完全削除 (Phase 2-B 完了後)

Refs #1186
